+++ CGraphicsLibrary/Matrix.h
@@ -9,28 +9,28 @@ enum class Axis {
 	Z = 'Z'
 };
 
-// 4 x 1 floating point vector
+// 4 x 1 doubleing point vector
 struct vec4 {
-	float x;
-	float y;
-	float z;
-	float w;
+	double x;
+	double y;
+	double z;
+	double w;
 	unsigned int color;
 
 	vec4();
-	vec4(float x, float y, float z, float w = 1.0f);
+	vec4(double x, double y, double z, double w = 1.0);
     void normalize();
     void hgDivision();
-    float dot(const vec4& rhs) const;
+    double dot(const vec4& rhs) const;
     /* Computes cross product between the difference of the three vertices passed in:
      * Performs V1 x V2
      */
     vec4 cross(const vec4& V2) const;
-    float length();
+    double length();
 	void print();
 
-	float operator[](int i) const;
-	vec4 operator*(float scalar) const;
+	double operator[](int i) const;
+	vec4 operator*(double scalar) const;
     vec4 operator*(const mat4& rhs) const;
     vec4 operator-(const vec4& p2) const;
     vec4 operator+(const vec4& p2) const;
@@ -62,7 +62,7 @@ enum class MAT_TYPE
 
 // 4 x 4 matrix
 struct mat4 {
-	float matrix[4][4];
+	double matrix[4][4];
 
 	// Init to identity
 	mat4();
@@ -70,9 +70,9 @@ struct mat4 {
     // Init to projection matrix
     mat4(MAT_TYPE type);
 
-	void scale(float x, float y, float z);
-	void translate(float x, float y, float z);
-	void rotate(float angle, Axis axis);
+	void scale(double x, double y, double z);
+	void translate(double x, double y, double z);
+	void rotate(double angle, Axis axis);
 
     void random(int min, int max);
     mat4 inverse();
